fix: SonarQube quick fixes batch 3 (S3490, S1854, S1155, S1709, S1905, S3642, S5028)
- S3490: WiFiConfigManager empty constructor -> = default
- S1854: Remove dead store in SequenceExecutor.cpp
- S1155: Use isEmpty() instead of length()>0 in APIRoutes.cpp
- S1709: Add explicit to MutexGuard ctor and operator bool
- S1905: Remove 3 redundant casts (FileSystem, OscillationController)
- S3642: enum class for NetworkMode and WatchdogState
- S5028: #define -> constexpr (WiFiConfigManager.h, SequenceTableManager.h)

+++ src/core/filesystem/FileSystem.cpp
@@ -116,7 +116,7 @@ String FileSystem::readFileAsString(const String& path, size_t maxSize) {
   if (!file) return "";
 
   // Limit read size for safety
-  size_t readSize = min((size_t)file.size(), maxSize);
+  size_t readSize = min(file.size(), maxSize);
   
   // Bulk read for performance
   char* buf = new (std::nothrow) char[readSize + 1];
